+++ rbac/management/utils.py
@@ -18,9 +18,11 @@
 import logging
 import os
 import uuid
+from contextlib import contextmanager
 from typing import Optional, TypedDict
 from uuid import UUID
 
+import grpc
 from django.conf import settings
 from django.core.exceptions import PermissionDenied
 from django.utils.translation import gettext as _
@@ -465,3 +467,10 @@ def flatten_validation_error(e: ValidationError):
         return [("__all__", str(msg)) for msg in e.messages]
     else:
         return [("__all__", str(e))]
+
+
+@contextmanager
+def create_client_channel(addr):
+    """Create secure channel for grpc requests."""
+    secure_channel = grpc.insecure_channel(addr)
+    yield secure_channel
